refactor: refactor categoty-navigation
