+++ src/feature/catalog/categories/category-path.tsx
@@ -13,3 +13,8 @@ export const getFullPath = (
 
   return path;
 };
+
+export const getCategoryPath = (categoryId: string, categories: BasicCategory[]) => {
+  const category = categories.find((c) => c.id === categoryId);
+  return category ? getFullPath(category, categories) : [];
+};
