+++ src/feature/catalog/categories/use-category-ui.ts
@@ -1,0 +1,11 @@
+import { useState } from 'react';
+
+export const useCategoryUI = () => {
+  const [openedSubmenus, setOpenedSubmenus] = useState<Record<string, boolean>>({});
+  const toggleSubmenu = (id: string) => setOpenedSubmenus((prev) => ({ ...prev, [id]: !prev[id] }));
+
+  return {
+    openedSubmenus,
+    toggleSubmenu,
+  };
+};
